+++ changelog.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.2] - 2025-02-24
+
+### Added
+- Use orjson for loading and dumping at rest end-point using
+  `PumpwoodJSONRenderer` and `PumpwoodJSONParser`.
+
+
+### Changed
+- No changes.
+
+### Removed
+- No Removes
+
 ## [1.5.X] - 2025-02-24
 
 ### Added